Feat: implement financial object factory
Refactored and reworked financial objects, now Revenue, Cost, ValueIncrease, ValueDecrease classes all implement the same IFinancialObject interface as they all use the same properties. This enabled the creation of a single Factory method that creates instances of these classes based on the properties of the received FinancialInputDTO. This signifficantly reduced the amount of tests and code needed.